fastapi/app/requirements.txt
fastapi
uvicorn
python-dotenv
configparser
mysql-connector-python
pymysql